While we're in here, add a setter for message.
This facilitates users having control over both the value and message of
their responses, if they want to update independently of the various
error/success functions. Strongly recommended to go through those, but
if a user really doesn't want the formatting of their message, don't
force them to have it.